Refactored the constraints module
Allows for easier implementations / reusability / extensions
