+++ cashocs/__init__.py
@@ -25,14 +25,11 @@ optimization problems, in particular, shape optimization and optimal control pro
 from . import verification
 from ._constraints.constraints import EqualityConstraint, InequalityConstraint
 
-from ._constraints.control_solvers import (
-    AugmentedLagrangianOptimalControlProblem,
-    QuadraticPenaltyOptimalControlProblem,
+from ._constraints.constrained_problems import (
+    ConstrainedOptimalControlProblem,
+    ConstrainedShapeOptimizationProblem,
 )
-from ._constraints.shape_solvers import (
-    AugmentedLagrangianShapeOptimizationProblem,
-    QuadraticPenaltyShapeOptimizationProblem,
-)
+
 from ._loggers import LogLevel, set_log_level
 from ._optimal_control.optimal_control_problem import OptimalControlProblem
 from ._shape_optimization.shape_optimization_problem import ShapeOptimizationProblem
@@ -71,8 +68,8 @@ __all__ = [
     "create_dirichlet_bcs",
     "moreau_yosida_regularization",
     "verification",
-    "AugmentedLagrangianOptimalControlProblem",
-    "AugmentedLagrangianShapeOptimizationProblem",
-    "QuadraticPenaltyShapeOptimizationProblem",
-    "QuadraticPenaltyOptimalControlProblem",
+    "ConstrainedOptimalControlProblem",
+    "ConstrainedShapeOptimizationProblem",
+    "EqualityConstraint",
+    "InequalityConstraint",
 ]
